Split settings into 2 files, rename 'html' to 'templates'
