+++ application/setup.py
@@ -9,7 +9,7 @@ from cx_Freeze import setup, Executable
 
 python3_dll = str(Path(sys.executable).parent / "python3.dll")
 templates = str(
-    Path(__file__).resolve().parents[1] / "src" / "splitnotes2" / "ui" / "html"
+    Path(__file__).resolve().parents[1] / "src" / "splitnotes2" / "templates"
 )
 icon_file = str(Path(__file__).resolve().parents[1] / "resources" / "logo_alpha.ico")
 icon_png = str(Path(__file__).resolve().parents[1] / "resources" / "logo_alpha.png")
