+++ src/splitnotes2/templates/default.css
