+++ src/splitnotes2/templates/default.html
@@ -6,8 +6,8 @@
   <style>
     :root {
       font-size: {{ font_size }}px;
-      color: {{font_color}};
-      background-color: {{bg_color}};
+      color: {{ font_color }};
+      background-color: {{ bg_color }};
     }
     {{ css }}
   </style>
